#39 Move domain model
