+++ webminer/entities/domain_model.py
