+++ webminer/external_interfaces/flask_server/rest/arxiv_document.py
@@ -3,10 +3,10 @@
 import json
 from flask import Blueprint, request, Response
 
-from webminer.interface_adapters.rest_adapters import request_objects as req
-from webminer.interface_adapters import process_arxiv_request as uc
-from webminer.interface_adapters.rest_adapters import response_object as res
 from webminer.use_cases.request_arxiv import arxiv_repo as ar
+from webminer.interface_adapters import process_arxiv_request as uc
+from webminer.interface_adapters.rest_adapters import request_objects as req
+from webminer.interface_adapters.rest_adapters import response_object as res
 from webminer.interface_adapters.serializers.json import arxiv_document_serializer as ser
 
 blueprint = Blueprint("arxiv", __name__)
